Wire together the middle and the client
Make an implementation of the command invoker, which at the moment does
not do much except for executing the evaluationcommand.

Let TerminalEditor be a TerminalUserInterface, by making it implement
IDisplay in addition to IEditor.

Set the onCompleteHook from the Repl class, and invoke the hook with a
dummy evaluation result for testing.