Test TerminalUI history by injecting ConsoleReader
The ConsoleReader is injected using an @Provides method in the
ReplModule.

In the test, it is injected with a special ConsoleReader in which I
explicitly set "terminal support", so that it interprets control codes
like UP, DOWN and ENTER. This way I can test as if I type in a terminal.