Split ReplTest into ConsoleReplTest and a new ReplTest in core
Because Repl was moved to core, ReplTest should test the Repl
generically with mock editor and displays since there is no Terminal UI.

We then let ConsoleReplTest be an integration test with a Terminal UI
and simulated input.

ConsoleReplTest extends from ReplTest and only changes the test setup.
ReplTest was brought into scope by building test-jars in the core
pom.xml, and adding a dependency in the console client pom.xml on the
test jar.