Add Editor Services interface
Provides Editor services that directly invoke the FailableFunctions by
using the FunctionComposer. The Editor Services can be injected using
Guice, to the frontend and used in a similar way to the CommandInvoker.
The presence of a separate class for this, is that the exact result type
can be determined, while with IReplCommands this was not the case.

The Editor uses a strategy pattern to define its behaviour before and
after loading a language. This is necessary to provide a clean interface
to the client, who must receive its EditorServices object from the
start, even though a language is not loaded initially, so that it
doesn't need to do null checks all the time.
The strategy of the EditorServices will decide whether it can provide a
meaningful result or simply return an ExceptionResult.