Split FunctionComposer from CommandBuilder
The commandbuilder dealt with both the function compositions (in terms
of private methods) and building commands.

The composed functions need to be exposed to be called by editor
services directly, rather than via the IReplCommand invocation.
This is important because the IReplCommand loses the return type
information. The exact expected type is important to the client, who
must use that result to implement the editor feature.

This change adds a FunctionComposer, simply specifying how to compose
functions.

It can be instantiated in the same way as the CommandBuilder: using the
factory method in IFunctionFactory.

The commandBuilder uses the composer to build the commands with, rather
than doing it himself.